pyproject.toml
[tool.poetry]
name = "fastapi-cloud-logging"
version = "0.8.4"
description = "Cloud Logging For FastAPI"
authors = ["quoth <4wordextinguisher@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/quoth/fastapi-cloud-logging"
repository = "https://github.com/quoth/fastapi-cloud-logging"
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = ">=0.71"
google-cloud-logging = "~3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^22.1.0"
pytest = "^7.0.1"
pytest-mock = "^3.7.0"
importlib_metadata = { version = "^4.2", optional = true, extras = ["py37"] }
uvicorn = {version = "^0.17.6", extras = ["local"]}
loguru = {version = "^0.6.0", extras = ["local"]}

[tool.poetry.extras]
py37 = ["importlib_metadata"]
local = ["uvicorn", "loguru"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.flake8]
max-line-length = 120

[tool.isort]
line_length = 120

[tool.pytest.ini_options]
addopts = [
  "--strict-config",
  "--strict-markers",
]
xfail_strict = true
filterwarnings = [
    "error",
    # TODO: needed by asyncio in Python 3.9.7 https://bugs.python.org/issue45097, try to remove on 3.9.8
    'ignore:Call to deprecated create function FieldDescriptor',
    'ignore:Call to deprecated create function Descriptor',
    'ignore:Call to deprecated create function EnumDescriptor',
    'ignore:Call to deprecated create function EnumValueDescriptor',
    'ignore:Call to deprecated create function FileDescriptor',
    'ignore:Call to deprecated create function OneofDescriptor'
]